% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventRelation.R
\name{analyseEventRelations}
\alias{analyseEventRelations}
\title{Analyse Event Relations}
\usage{
analyseEventRelations(eventRelations)
}
\arguments{
\item{eventRelations}{data frame as returned by
\code{\link{getEventRelations}}}
}
\description{
Analyse Event Relations
}
