% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{eventDuration}
\alias{eventDuration}
\title{Event Duration}
\usage{
eventDuration(tBeg, tEnd, signalWidth)
}
\arguments{
\item{tBeg}{timestamps representing the event begins}

\item{tEnd}{timestamps representing the event ends}

\item{signalWidth}{see description in \code{\link{hsEvents}}}
}
\description{
Event Duration
}
