% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{eventLimits}
\alias{eventLimits}
\title{only tBeg and tEnd of events (possibly extended)}
\usage{
eventLimits(events, context = c(0, 0), absolute = FALSE)
}
\arguments{
\item{events}{events as returned by e.g. \code{\link{toEvents}}}

\item{context}{Vector of two elements giving the "context" before and after
the event to be plotted, in percentage of event duration. e.g. c(0.1, 0.2)
means that the time interval to be plotted starts 10 percent of the event 
duration before the event begin and ends 20 percent of the event duration 
after the end of the event.}

\item{absolute}{if TRUE, the context values are interpreted as absolute
values (seconds) instead of fractions of the event duration. Default: FALSE}
}
\value{
data frame with columns \emph{tBeg} and \emph{tEnd}, taken from 
  \emph{events} and possibly reduced (tBeg) and/or extended (tEnd) by a 
  fraction of the event duration (read from column \emph{dur} in 
  \emph{events}).
}
\description{
return only the limits \emph{tBeg} and \emph{tEnd} of the event, possibly 
  extended by a "context"
}
