% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{eventPauses}
\alias{eventPauses}
\title{Pauses between Events in Seconds}
\usage{
eventPauses(events, signalWidth = .getSignalWidth(events),
  timeDifferences = NULL)
}
\arguments{
\item{timeDifferences}{if time differences have been calculated beforehand,
these may be given here (in seconds)}
}
\value{
data frame with columns \emph{pBefore}, \emph{pEnd} with the pauses 
  between the given \emph{events}, in seconds
}
\description{
Pauses between Events in Seconds
}
