% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventRelation.R
\name{eventRelation}
\alias{eventRelation}
\title{Relations of Start/End Times of Events}
\usage{
eventRelation(event1, events2)
}
\arguments{
\item{event1}{data frame containing exactly one row representing the event to
which the event(s) in \emph{events2} is/are to be compared. Columns
\emph{tBeg} (begin of event), \emph{tEnd} (end of event) and \emph{event}
(event number/ID) are required.}

\item{events2}{data frame containing in rows the event(s) that are to be
compared to the event given in \emph{event1}. Columns \emph{tBeg} (begin of
event), \emph{tEnd} (end of event) and \emph{event} (event number/ID) are
required.}
}
\description{
Relations of Start/End Times of Events
}
