% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{eventsOnChange}
\alias{eventsOnChange}
\title{Changes in Value Vector to Events}
\usage{
eventsOnChange(x, numberOnly = FALSE, include.value = FALSE)
}
\arguments{
\item{x}{vector containing elements to be grouped into \dQuote{events}}

\item{numberOnly}{if TRUE, only the number of \dQuote{events} is returned
instead of a data frame containing first and last index of each
\dQuote{event}.}

\item{include.value}{if TRUE and \emph{numberOnly} is FALSE, the returned
data frame will contain a column \emph{value} containing the value that was
found in each index section between \emph{iBeg} and \emph{iEnd}.}
}
\value{
Per default (\emph{numberOnly} = FALSE) a data frame is returned with
  as many rows as \dQuote{events} were found in vector \emph{x}. As long as
  the value in \emph{x} does not change from one index to the next, it is 
  assumed to belong to the same event. If the value changes, a new event 
  begins. In the result data frame each event is represented by \emph{iBeg} 
  and \emph{iEnd} which are the indices of the first and last element, 
  respectively, in \emph{x} that build the event. If \emph{numberOnly} is 
  TRUE the number of \dQuote{events} is returned, that is one plus the number
  of changes in the value of \emph{x} from its first to its last element.
}
\description{
Creates \dQuote{events} from vector \emph{x} of values based on changes in 
the value of consecutive elements in \emph{x}.
}
\examples{
eventsOnChange(c(1,2,2,3,4,4,4,5))
  
# Ouput: list of five events, i.e. there are four changes of 
#        the value in the given vector.
#
#   iBeg iEnd
# 1    1    1
# 2    2    3
# 3    4    4
# 4    5    7
# 5    8    8
  
eventsOnChange(c(1, 2, 2, 3, 4, 4, 4, 5), numberOnly = TRUE) ## 5 (events)
  
}
\seealso{
\code{\link{hsEvents}}
}
