% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{eventToXLim}
\alias{eventToXLim}
\title{tBeg and tEnd of event to two-element vector}
\usage{
eventToXLim(event)
}
\arguments{
\item{event}{data frame with columns \emph{tBeg}, \emph{tEnd}}
}
\value{
vector of two elements: \emph{tBeg} and \emph{tEnd} from
  \emph{event}, both converted to UTC timezone
}
\description{
puts tBeg and tEnd of event into a vector of two POSIXct elements
}
