% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventHighLevel.R
\name{filterEventsWithStatistics}
\alias{filterEventsWithStatistics}
\title{Filter Events with Statistics}
\usage{
filterEventsWithStatistics(eventData, durationThreshold = 0,
  durationComparisonOperator = "gt", sumThreshold = 0,
  sumComparisonOperator = "gt")
}
\arguments{
\item{eventData}{eventData of one time series, as returned in one list
element by \code{\link{getEventsWithStatistics}}.}

\item{durationThreshold}{duration in seconds that needs to be exceeded
(durationComparisonOperator == "gt") or reached (durationComparisonOperator
== "ge") by the duration of the rain events. Default: 0}

\item{durationComparisonOperator}{Operator to be applied when comparing the
duration of the events with \emph{durationThreshold}. Must be one of "gt"
(greater than) or "ge" greater than or equal. Default: "gt"}

\item{sumThreshold}{value that needs to be exceeded (sumComparisonOperator ==
"gt") or reached (sumComparisonOperator == "ge") by the 'sum' of values
within the events. Default: 0}

\item{sumComparisonOperator}{Operator to be applied when comparing the 'sum'
of values within the events with \emph{sumThreshold}. Must be one of "gt"
(greater than) or "ge" greater than or equal. Default: "gt"}
}
\description{
Filter Events with Statistics
}
