% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventPlot.R
\name{ganttPlotEvents}
\alias{ganttPlotEvents}
\title{Gantt-like Diagram to plot Event's Time Extension}
\usage{
ganttPlotEvents(
  events,
  add = FALSE,
  y1 = 1,
  y2 = y1 + 1,
  xlim = NULL,
  ylim = c(min(y1), max(y2)),
  col = "black",
  density = 5,
  showLabels = TRUE,
  eventLabels = rownames(events),
  yLabel = (y1 + y2)/2,
  type = "rectangle",
  title = "",
  leftMargin = 0.2,
  xlab = "Time",
  cex = 0.8,
  indicate = NULL,
  indicationColour = "red",
  bandheight = 0.1,
  alternating = FALSE,
  adj = 0.5,
  ...
)
}
\arguments{
\item{events}{event list as retrieved by \code{\link{hsEvents}}. Required
columns: \emph{tBeg} (begin of event) and \emph{tEnd} (end of event), both
of class POSIXt}

\item{add}{if TRUE, the event boxes are added to the current plot, otherwise
a new plot is generated}

\item{y1}{lower coordinates of the event boxes}

\item{y2}{upper coordinates of the event boxes}

\item{xlim}{x limits. If NULL (default) the limits will be chosen so that all
events fit into the plot}

\item{ylim}{y limits}

\item{col}{colour of shading lines}

\item{density}{density of shading lines}

\item{showLabels}{if TRUE, the event boxes are labelled with the row names of
the events}

\item{eventLabels}{labels to be given to the events. Default:
rownames(\emph{events})}

\item{yLabel}{y-position of labels, if labels are to be shown}

\item{type}{one of c("rectange", "vertical")}

\item{title}{title to be plotted left of event rectangles}

\item{leftMargin}{left margin (where title is printed) as fraction of the
range of the total time interval spanned by the events}

\item{xlab}{x axis label}

\item{cex}{character expansion factor}

\item{indicate}{indices of events to be indicated in a different color
(indicationColuor)}

\item{indicationColour}{colour to be used for indication, default: "red" 
extension factor for labels (event numbers)}

\item{bandheight}{passed to \code{addLabels}}

\item{alternating}{passed to \code{addLabels}}

\item{adj}{passed to \code{text} plotting the event labels}

\item{\dots}{further arguments passed to rect or segments}
}
\description{
Gantt-like Diagram to plot Event's Time Extension
}
