% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventHighLevel.R
\name{getAndFilterEventsWithStatistics}
\alias{getAndFilterEventsWithStatistics}
\title{Get and filter Events with Statistics}
\usage{
getAndFilterEventsWithStatistics(rainData, seriesName, eventSeparationTime,
  signalThreshold = 0, durationThreshold = 1, sumThreshold = 0,
  signalComparisonOperator = "gt", durationComparisonOperator = "gt",
  sumComparisonOperator = "gt", signalWidth = NA)
}
\description{
Get and filter Events with Statistics
}
