% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventHighLevel.R
\name{getEvents}
\alias{getEvents}
\title{Get Events}
\usage{
getEvents(
  rainData,
  seriesName,
  signalThreshold = 0,
  signalComparisonOperator = "gt",
  eventSeparationTime = 6 * 3600,
  eventSeparationOperator = "gt",
  signalWidth = NA,
  column.time = names(rainData)[1]
)
}
\arguments{
\item{rainData}{data frame with time stamps in the first column and rain
heights (or intensities) in the remaining columns}

\item{seriesName}{Column name in rainData representing the time series to be
analysed.}

\item{signalThreshold}{Value that needs to be exceeded
(signalComparisonOperator == "gt") or reached (signalComparisonOperator ==
"ge") by the rain heights (or intensities) in order to be counted as a
"signal". Default: 0}

\item{signalComparisonOperator}{Operator to be applied when comparing rain
values with signalThreshold. Must be one of "gt" (greater than) or "ge"
greater than or equal. Default: "gt".}

\item{eventSeparationTime}{Time difference in seconds that needs to be
exceeded (eventSeparationOperator == "gt") or reached 
(eventSeparationOperator == "ge") by two consecutive signals in order to
let the signals belong to two distinct events. Otherwise the signals are
assumed to belong to one and the same event. Default: 6*3600 = six hours.}

\item{eventSeparationOperator}{Operator to be applied when comparing the time
differences between consecutive signals with the eventSeparationTime. Must
be one of "gt" (greater than) or "ge" greater than or equal. Default: "gt".}

\item{signalWidth}{signal width (= length of the time interval represented by
one row in \code{rainData}) in seconds}

\item{column.time}{name of the column containing the time. Default: Name of
the first column}
}
\description{
Get Events
}
