% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventHighLevel.R
\name{getEventStatistics}
\alias{getEventStatistics}
\title{Get Event Statistics}
\usage{
getEventStatistics(dataFrame, seriesName, events, functions = c("sum", "mean",
  "min", "max", "number.na"), eventNumbers = 1:nrow(events))
}
\arguments{
\item{functions}{define statistical functions}

\item{eventNumbers}{vector of same length as \emph{events} has rows, giving
the numbers that identify the events. Default: 1:nrow(\emph{events})}
}
\value{
data frame with event number in first column \emph{event} and
  statistical values in further columns.
}
\description{
Get Event Statistics
}
