% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventHighLevel.R
\name{getEventsWithStatistics}
\alias{getEventsWithStatistics}
\title{Get Events with Statistics}
\usage{
getEventsWithStatistics(
  rainData,
  seriesName,
  eventSeparationTime,
  signalThreshold = 0,
  signalComparisonOperator = "gt",
  eventSeparationOperator = "gt",
  functions = c("sum", "mean", "min", "max", "number.na", "length"),
  signalWidth = NA
)
}
\arguments{
\item{rainData}{data frame with time stamps in the first column and rain
heights (or intensities) in the remaining columns}

\item{seriesName}{Column name in rainData representing the time series to be
analysed.}

\item{eventSeparationTime}{passed to \code{\link{getEvents}}}

\item{signalThreshold}{value that needs to be exceeded
(signalComparisonOperator == "gt") or reached (signalComparisonOperator ==
"ge") by the rain heights (or intensities) in order to be counted as a
"signal". Default: 0}

\item{signalComparisonOperator}{Operator to be applied when comparing rain
values with signalThreshold. Must be one of "gt" (greater than) or "ge"
greater than or equal. Default: "gt"}

\item{eventSeparationOperator}{passed to \code{\link{getEvents}}}

\item{functions}{passed to \code{\link{getEventStatistics}}}

\item{signalWidth}{passed to \code{\link{getEvents}}}
}
\description{
Get Events with Statistics
}
