% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventHighLevel.R
\name{getEventsWithStatisticsForMultipleSeries}
\alias{getEventsWithStatisticsForMultipleSeries}
\title{Get Events with Statistics for multiple Series}
\usage{
getEventsWithStatisticsForMultipleSeries(
  rainData,
  eventSeparationTime,
  signalWidth = kwb.datetime::getTimestepInSeconds(timestamps = rainData[, 1]),
  signalThreshold = 0,
  signalComparisonOperator = "gt"
)
}
\arguments{
\item{rainData}{data frame with time stamps in the first column and rain
heights (or intensities) in the remaining columns}

\item{eventSeparationTime}{\dQuote{event separation time} in seconds. Maximal
allowed time difference between two consecutive timestamps within the same
event.}

\item{signalWidth}{\dQuote{signal width} in seconds. Length of time interval
that one timestamp is representing, e.g. \eqn{5*60 = 300} if each timestamp
respresents a time interval of five minutes (as e.g. a time series is 
recorded on a five minute time scale). This parameter is needed to 
calculate event durations.}

\item{signalThreshold}{value that needs to be exceeded
(signalComparisonOperator == "gt") or reached (signalComparisonOperator ==
"ge") by the rain heights (or intensities) in order to be counted as a
"signal". Default: 0}

\item{signalComparisonOperator}{Operator to be applied when comparing rain
values with signalThreshold. Must be one of "gt" (greater than) or "ge"
greater than or equal. Default: "gt"}
}
\description{
Get Events with Statistics for multiple Series
}
