% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventRelation.R
\name{getParallelEventNotEndingAfter}
\alias{getParallelEventNotEndingAfter}
\title{"Parallel" Events}
\usage{
getParallelEventNotEndingAfter(events1, events2, eventRelations,
  extended = FALSE)
}
\arguments{
\item{events1}{data frame containing the reference events, e.g. discharge
events}

\item{events2}{parallel events, e.g. rain events}

\item{eventRelations}{event relations as returned by
\code{\link{getEventRelations}}}
}
\description{
Calculates event borders (event begin, event end) considering "parallel" 
events2. The returned results not ending after events1. For each event
\emph{E} in \emph{events1} (defined by event number, event begin and event
end time), this function first identifies the "partner" events E2,i from
\emph{events2} that lie within \emph{E} or have an intersecton with \emph{E}.
There may be no, one ore more than one "partner" events.
}
