% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMerge.R
\name{getParallelEventsInfo}
\alias{getParallelEventsInfo}
\title{Information on Events in parallel}
\usage{
getParallelEventsInfo(eventLists, version = 1)
}
\arguments{
\item{eventLists}{list of data frames, each of which represents a list of
events as e.g. generated by \code{\link{hsEvents}}}
}
\value{
data frame with timestamps in the first column indicating any begin
  or end of any event within \emph{eventLists} and columns for each element
  of \emph{eventLists}, containing event numbers. If you go along one row you
  can find the events that occur in parallel.
}
\description{
Information on Events in parallel
}
