% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{getXLimFromEventLists}
\alias{getXLimFromEventLists}
\title{overall first begin and last end of events}
\usage{
getXLimFromEventLists(eventLists)
}
\arguments{
\item{eventLists}{list of data frames containing events (containing columns
\emph{tBeg}, \emph{tBeg}, as returned by \code{\link{hsEvents}})}
}
\value{
vector of two elements: the first begin (minimum of tBeg) and the
  last end (maximum of tEnd), found in any of the event data frames given in 
  \emph{eventLists}
}
\description{
minimum tBeg and maximum tEnd found in event lists
}
