% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{hsEvents}
\alias{hsEvents}
\title{Timestamp differences to events}
\usage{
hsEvents(tseries, evtSepTime, signalWidth, tUnit = "s", pause = TRUE,
  evtSepOp = "gt", dbg = FALSE, check.sorting = FALSE)
}
\arguments{
\item{tseries}{vector containing a sorted list of timestamps.}

\item{evtSepTime}{\dQuote{event separation time} in seconds. Maximal allowed
time difference between two consecutive timestamps within the same event.}

\item{signalWidth}{\dQuote{signal width} in seconds. Length of time interval
that one timestamp is representing, e.g. \eqn{5*60 = 300} if each timestamp
respresents a time interval of five minutes (as e.g. a time series is
recorded on a five minute time scale). This parameter is needed to
calculate event durations.}

\item{tUnit}{time unit of event duration and event pauses}

\item{pause}{if TRUE, pauses before and after the events are calculated}

\item{evtSepOp}{event separation operator, either "gt" or "ge". If
\emph{evtSepOp} = "gt" (default) events are separated on time differences
between two consecutive timestamps that are \emph{g}reater \emph{t}han
\emph{evtSepTime}. If \emph{evtSepOp} = "ge" events are separated on time
differences between two consecutive timestamps that are \emph{g}reater than
or \emph{e}qual to \emph{evtSepTime}.}

\item{dbg}{if TRUE, debug messages are shown.}
}
\value{
data frame with columns \emph{iBeg} and \emph{iEnd} indicating first
  and last index of the event in the \emph{tseries} vector, \emph{tBeg} and 
  \emph{tEnd} indicating first and last timestamp of the event and \emph{dur}
  indicating the event duration in seconds.
}
\description{
Creates events from vector \emph{tseries} of timestamps based on time 
differences between consecutive timestamps in \emph{tseries}.
}
\seealso{
\code{\link{hsEventsOnChange}}
}
