% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{hsEventsOnChange}
\alias{hsEventsOnChange}
\title{Deprecated. Use eventsOnChange() instead.}
\usage{
hsEventsOnChange(...)
}
\arguments{
\item{\dots}{passed to \code{\link{eventsOnChange}}}
}
\description{
Deprecated. Use eventsOnChange() instead.
}
