% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{hsEventsToUnit}
\alias{hsEventsToUnit}
\title{Time unit conversion for events}
\usage{
hsEventsToUnit(evts, tUnit)
}
\arguments{
\item{evts}{data frame representing events as provided by hsEvents}

\item{tUnit}{time unit to which durations and pauses shall be converted.}
}
\value{
data frame containing events with durations (and pauses) given in the
  new time unit.
}
\description{
Converts event durations and pauses before and after events to the requested 
time unit. The time unit will be stored in the attribute \dQuote{tUnit} of
the returned data frame.
}
