% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{hsGetEvent}
\alias{hsGetEvent}
\title{Get Sub-Timeseries belonging to Event(s)}
\usage{
hsGetEvent(tSeries, events, evtnums, useIndex = FALSE)
}
\arguments{
\item{tSeries}{data frame representing time series with first column holding
the timestamp}

\item{events}{event information as returned by \code{\link{hsEvents}}}

\item{evtnums}{vector of event numbers to be selected}

\item{useIndex}{if TRUE, \emph{tSeries} is filtered by comparing the real row
number in \emph{tSeries} with the begin and end indices given in columns 
\emph{iBeg} and \emph{iEnd} of \emph{events}. If FALSE, tSeries is filtered
by comparing the timestamps in tSeries with the begin and end timestamps
given in columns \emph{tBeg} and \emph{tEnd} of \emph{events}. Defaults to
TRUE if \emph{events} contains columns \emph{iBeg} and \emph{iEnd}}
}
\value{
rows of \emph{tSeries} belonging to the event numbers listed in 
  \emph{evtnums}
}
\description{
Get Sub-Timeseries belonging to Event(s)
}
