% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventJoin.R
\name{hsJoinEvents}
\alias{hsJoinEvents}
\title{Join events in event list}
\usage{
hsJoinEvents(evts, ..., renumber = TRUE, dbg = FALSE)
}
\arguments{
\item{evts}{data frame containing events as provided by e.g. hsEvents}

\item{\dots}{numeric vectors containing the event numbers to be joined, e.g. 
5:10, 15:20 will join events 5 to 10 and 15 to 20 to one event in each case}

\item{renumber}{if TRUE, rows in result data frame are renumbered from one to
number of rows.}

\item{dbg}{if \code{TRUE}, debug messages are shown.}
}
\value{
A data frame with fields \emph{tBeg}, \emph{tEnd}, \emph{dur}
  containing the times of event begin and event end and the event duration in
  seconds, respectively. The event duration is the difference between end and
  begin of the event plus the time period that one timestamp represents 
  (signal width).
}
\description{
Join consecutive events in event list \emph{evts}. The result of joining two
events A and B is a event with begin time of A and end time of B.
}
