% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMerge.R
\name{hsMergeEvents}
\alias{hsMergeEvents}
\title{Merge two Event Lists}
\usage{
hsMergeEvents(events1, events2, renumber = TRUE, dbg = FALSE)
}
\arguments{
\item{events1}{data frame containing events as provided by e.g. hsEvents}

\item{events2}{data frame containing events as provided by e.g. hsEvents}

\item{renumber}{if TRUE, rows in result data frame are renumbered from one to
number of rows.}
}
\value{
data frame with fields \emph{tBeg}, \emph{tEnd}, \emph{dur}
  containing the times of event begin and event end and the event duration in
  seconds, respectively. The event duration is the difference between end and
  begin of the event plus the time period that one timestamp represents 
  (signal width).
}
\description{
Events in data frames \emph{events1} and \emph{events2} are merged in such a 
way that overlapping events are combined to one event and events that are 
fully contained in other events are discarded.
}
