% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{hsSigWidth}
\alias{hsSigWidth}
\title{Find signal width in event list}
\usage{
hsSigWidth(evts, dbg = FALSE)
}
\arguments{
\item{evts}{data frame containing events (as e.g. provided by hsEvents)}

\item{dbg}{if \code{TRUE}, debug messages are shown.}
}
\value{
signal width in seconds
}
\description{
Calculates signal width that was applied in event list \emph{evts}
}
