% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{indicesOfEventsContainingEvent}
\alias{indicesOfEventsContainingEvent}
\title{indicesOfEventsContainingEvent}
\usage{
indicesOfEventsContainingEvent(events, event)
}
\arguments{
\item{events}{data frame with columns \emph{tBeg}, \emph{tEnd}}

\item{event}{data frame of one row with columns \emph{tBeg}, \emph{tEnd}}
}
\value{
vector of indices representing the positions of the events in 
  \emph{events} in which \emph{event} is fully contained
}
\description{
indicesOfEventsContainingEvent
}
