% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{invertedEvents}
\alias{invertedEvents}
\title{Events to Gaps between Events}
\usage{
invertedEvents(events)
}
\arguments{
\item{events}{data frame with columns \emph{tBeg} (begin of event) and
\emph{tEnd} (end of event), representing events}
}
\description{
"inverted" events: gaps between ends of events and begins of next events
}
