% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMerge.R
\name{mergeAllEvents}
\alias{mergeAllEvents}
\title{Merge all Events}
\usage{
mergeAllEvents(eventList, dbg = TRUE)
}
\arguments{
\item{eventList}{list of data frames, each of which represents a list of
events as e.g. generated by \code{\link{hsEvents}}}

\item{dbg}{if TRUE, debug messages are shown}
}
\description{
'merge' all events in a list of event lists
}
