% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventRelation.R
\name{overlapping}
\alias{overlapping}
\title{Are there overlapping Events?}
\usage{
overlapping(events)
}
\arguments{
\item{events}{event list as returned by \code{\link{hsEvents}}}
}
\value{
TRUE if there are overlapping events, otherwise FALSE
}
\description{
Check event list for overlaps (begin of one event before the end of a 
  previous event)
}
\examples{
events <- kwb.event::exampleEvents()
  
# The example events do not overlap
overlapping(events)
  
# The order of the events (here reverse order) does not matter
overlapping(events[nrow(events):1, ])
  
# Put the begin of the second event before the end of the last event
events$tBeg[2] <- mean(c(events$tBeg[1], events$tEnd[1]))
  
# Now there are overlapping events!
overlapping(events)
  
}
