% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventPlot.R
\name{plotEventInfo}
\alias{plotEventInfo}
\title{Plot Event Info}
\usage{
plotEventInfo(eventInfo)
}
\arguments{
\item{eventInfo}{as returned by \emph{getParallelEventsInfo}, with first
columns (timestamps) removed}
}
\description{
Plot Event Info
}
