% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventPlot.R
\name{plotMergedEventInfoForValidation}
\alias{plotMergedEventInfoForValidation}
\title{Plot merged Event Info for Validation}
\usage{
plotMergedEventInfoForValidation(mergedEvents)
}
\arguments{
\item{mergedEvents}{data frame containing information about merged events,
i.e. containing columns \code{tBeg.event1}, \code{tEnd.event1}, 
\code{tBeg.event2first}, \code{tEnd.event2last}, \code{tBeg.merged}, 
\code{tEnd.merged}}
}
\description{
Plot merged Event Info for Validation
}
