% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventIO.R
\name{readEventFilesFromDirectory}
\alias{readEventFilesFromDirectory}
\title{Read Event Files from Directory}
\usage{
readEventFilesFromDirectory(
  event.dir,
  prefix = "events",
  timezone = "UTC",
  ...
)
}
\arguments{
\item{event.dir}{full path to directory containing event definition files}

\item{prefix}{prefix of file names to be searched for}

\item{timezone}{timezone to which the timestamps are to be converted.
Default: "UTC"}

\item{\dots}{arguments passed to \code{\link{readEventsFromFile}}}
}
\description{
Read event definitions from files "events_*.txt" in event.dir
}
