% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventIO.R
\name{readEventsFromFile}
\alias{readEventsFromFile}
\title{Read Event Limits from File}
\usage{
readEventsFromFile(file, timezone = "UTC", header = FALSE)
}
\arguments{
\item{file}{full path to file containing the event definitions}

\item{timezone}{timezone to which the timestamps are to be converted.
Default: "UTC"}

\item{header}{TRUE if the file contains a header line (first
non-comment-line). Default: FALSE. If the file contains a header line, it
must contain the column captions "tBeg" and "tEnd" (begin and end
timestamps of the event).}
}
\value{
data frame with columns \emph{tBeg} and \emph{tEnd} (POSIXct)
}
\description{
Read Event Limits from File
}
