% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{toEvents}
\alias{toEvents}
\title{Convert to Data Frame of Events}
\usage{
toEvents(events, signalWidth = .getSignalWidth(events, default = NA),
  timeUnit = "s", pause = TRUE, timeDifferences = NULL)
}
\arguments{
\item{events}{data frame with columns \emph{tBeg} (event begins) and
\emph{tEnd} (event ends)}

\item{signalWidth}{\dQuote{signal width} in seconds. Length of time interval
that one timestamp is representing, e.g. \eqn{5*60 = 300} if each timestamp
respresents a time interval of five minutes (as e.g. a time series is
recorded on a five minute time scale). This parameter is needed to
calculate event durations.}

\item{timeUnit}{time unit of event duration and event pauses}

\item{pause}{if TRUE, pauses before and after the events are calculated}

\item{timeDifferences}{if time differences have been calculated beforehand,
these may be given here (in seconds)}
}
\description{
Convert to Data Frame of Events
}
