% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventHighLevel.R
\name{validateEventFunctionArguments}
\alias{validateEventFunctionArguments}
\title{Validate Event Function Arguments}
\usage{
validateEventFunctionArguments(...)
}
\arguments{
\item{\dots}{arguments, given as \code{key = value} pairs, to be checked for 
validity}
}
\description{
Validate Event Function Arguments
}
