% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventHighLevel.R
\name{whichAboveThreshold}
\alias{whichAboveThreshold}
\title{Which Values are above a Threshold?}
\usage{
whichAboveThreshold(values, threshold, comparisonOperator)
}
\description{
Which Values are above a Threshold?
}
