% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{eventDuration}
\alias{eventDuration}
\title{Event Duration}
\usage{
eventDuration(tBeg, tEnd, signalWidth)
}
\arguments{
\item{tBeg}{timestamps representing the event begins}

\item{tEnd}{timestamps representing the event ends}

\item{signalWidth}{\dQuote{signal width} in seconds. Length of time interval
that one timestamp is representing, e.g. \eqn{5*60 = 300} if each timestamp
respresents a time interval of five minutes (as e.g. a time series is
recorded on a five minute time scale). This parameter is needed to
calculate event durations.}
}
\description{
Event Duration
}
