% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventPlot.R
\name{ganttPlotEventLists}
\alias{ganttPlotEventLists}
\title{Gantt Plot of Event Lists}
\usage{
ganttPlotEventLists(eventLists, margin.top = 0.8, time.format = NULL,
  n.xticks = 10, showLabels = TRUE, ...)
}
\arguments{
\item{eventLists}{list of data frames containing events (containing columns
\emph{tBeg}, \emph{tBeg}, as returned by \code{\link{hsEvents}})}

\item{\dots}{further arguments passed to ganttPlotEvents}
}
\description{
Plot event lists, one above the other
}
