% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventHighLevel.R
\name{getAndFilterEventsWithStatistics}
\alias{getAndFilterEventsWithStatistics}
\title{Get and filter Events with Statistics}
\usage{
getAndFilterEventsWithStatistics(
  rainData,
  seriesName,
  eventSeparationTime,
  signalThreshold = 0,
  durationThreshold = 1,
  sumThreshold = 0,
  signalComparisonOperator = "gt",
  durationComparisonOperator = "gt",
  sumComparisonOperator = "gt",
  signalWidth = NA
)
}
\arguments{
\item{rainData}{passed to \code{\link{getEventsWithStatistics}}}

\item{seriesName}{passed to \code{\link{getEventsWithStatistics}}}

\item{eventSeparationTime}{passed to \code{\link{getEventsWithStatistics}}}

\item{signalThreshold}{passed to \code{\link{getEventsWithStatistics}}}

\item{durationThreshold}{passed to \code{\link{filterEventsWithStatistics}}}

\item{sumThreshold}{passed to \code{\link{filterEventsWithStatistics}}}

\item{signalComparisonOperator}{passed to \code{\link{getEventsWithStatistics}}}

\item{durationComparisonOperator}{\code{\link{filterEventsWithStatistics}}}

\item{sumComparisonOperator}{\code{\link{filterEventsWithStatistics}}}

\item{signalWidth}{passed to \code{\link{getEventsWithStatistics}}}
}
\description{
Get and filter Events with Statistics
}
