% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventPlot.R
\name{plotEventProperty1VersusEventProperty2}
\alias{plotEventProperty1VersusEventProperty2}
\title{Plot Event Property 1 versus Event Property 2}
\usage{
plotEventProperty1VersusEventProperty2(events, propertyName1, propertyName2,
  eventNumbers = events$eventNumber, xlab = propertyName1,
  ylab = propertyName2, cex = 0.7, ...)
}
\arguments{
\item{events}{data frame with at least two columns named as given in 
\emph{propertyName1} and \emph{propertyName2}}

\item{eventNumbers}{vector of event numbers used for labelling. Default:
rownames of \emph{events}}

\item{xlab}{default: propertyName1}

\item{ylab}{default: propertyName2}
}
\description{
Plot Event Property 1 versus Event Property 2
}
