% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventPlot.R
\name{plotMergedEventInfoForValidation}
\alias{plotMergedEventInfoForValidation}
\title{Plot merged Event Info for Validation}
\usage{
plotMergedEventInfoForValidation(mergedEvents)
}
\description{
Plot merged Event Info for Validation
}
