% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{renumberEvents}
\alias{renumberEvents}
\title{renumberEvents}
\usage{
renumberEvents(events)
}
\arguments{
\item{events}{event information as returned by \code{\link{hsEvents}}}
}
\value{
data frame with (additional) column \emph{event}
}
\description{
add event number (= real row number) in column \emph{event}
}
