% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventMain.R
\name{toEvents}
\alias{toEvents}
\title{Convert to Data Frame of Events}
\usage{
toEvents(events, signalWidth = .getSignalWidth(events, default = NA),
  timeUnit = "s", pause = TRUE, timeDifferences = NULL)
}
\arguments{
\item{events}{data frame with columns \emph{tBeg} (event begins) and
\emph{tEnd} (event ends)}

\item{signalWidth}{see description in \code{\link{hsEvents}}}

\item{timeUnit}{time unit of event duration and event pauses}

\item{pause}{if TRUE, pauses before and after the events are calculated}

\item{timeDifferences}{if time differences have been calculated beforehand,
these may be given here (in seconds)}
}
\description{
Convert to Data Frame of Events
}
