% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventHighLevel.R
\name{validateEventFunctionArguments}
\alias{validateEventFunctionArguments}
\title{Validate Event Function Arguments}
\usage{
validateEventFunctionArguments(...)
}
\description{
Validate Event Function Arguments
}
