% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventHighLevel.R
\name{whichAboveThreshold}
\alias{whichAboveThreshold}
\title{Which Values are above a Threshold?}
\usage{
whichAboveThreshold(values, threshold, comparisonOperator)
}
\arguments{
\item{values}{numeric vector of values to be compared with the 
\code{threshold}}

\item{threshold}{numeric value against which to check the \code{values}}

\item{comparisonOperator}{if "gt" it is checked whether the \code{values} 
are \emph{greater than} the \code{threshold}, otherwise it is checked
whether the \code{values} are \code{greater than or equal to} the
\code{threshold}}
}
\description{
Which Values are above a Threshold?
}
